local-companion/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pyautogui>=0.9.54
pytesseract>=0.3.10
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pydantic>=2.0.0
python-multipart>=0.0.6